+++ eqcorrscan/tests/correlate_test.py
@@ -9,6 +9,7 @@ import numpy as np
 import pytest
 from multiprocessing import cpu_count
 from obspy import Trace, Stream, read
+from obspy.core.util import AttribDict
 from scipy.fftpack import next_fast_len
 
 import eqcorrscan.utils.correlate as corr
@@ -98,6 +99,16 @@ def generate_multichannel_templates():
     return templates
 
 
+def generate_weighted_multichannel_templates():
+    random.seed(42)
+    templates = generate_multichannel_templates()
+    for template in templates:
+        for tr in template:
+            tr.stats.extra = AttribDict()
+            tr.stats.extra.update({"weight": random.random()})
+    return templates
+
+
 def read_gappy_real_template():
     return [read(join(pytest.test_data_path, "DUWZ_template.ms"))]
 
@@ -164,6 +175,8 @@ def print_class_name(request):
 # array fixtures
 
 starting_index = 500
+GLOBAL_WEIGHT = 0.5  # Weight applied to array ccs funcs
+# global to allow comparison to unweighted
 
 
 @pytest.fixture(scope='module')
@@ -192,6 +205,29 @@ def pads(array_template):
     return an array of zeros for padding, matching templates len.
     """
     return np.zeros(array_template.shape[0], dtype=int)
+
+
+@pytest.fixture(scope='module')
+def array_ccs_weighted(array_template, array_stream, pads):
+    """ Use each function stored in the normxcorr cache to correlate the
+     templates and arrays, return a dict with keys as func names and values
+     as the cc calculated by said function"""
+    out = {}
+    weights = np.ones(array_template.shape[0]) * GLOBAL_WEIGHT
+    for name in list(corr.XCORR_FUNCS_ORIGINAL.keys()):
+        func = corr.get_array_xcorr(name)
+        print("Running %s" % name)
+        cc, _ = time_func(
+            func, name, array_template, array_stream, pads, weights)
+        out[name] = cc
+        if "fftw" in name:
+            print("Running fixed len fft")
+            fft_len = next_fast_len(
+                max(len(array_stream) // 4, len(array_template)))
+            cc, _ = time_func(func, name, array_template, array_stream, pads,
+                              weights, fft_len=fft_len)
+            out[name + "_fixed_len"] = cc
+    return out
 
 
 @pytest.fixture(scope='module')
@@ -252,6 +288,12 @@ def multichannel_templates():
 
 
 @pytest.fixture(scope='module')
+def weighted_multichannel_templates():
+    """ create weighted multichannel templates """
+    return generate_weighted_multichannel_templates()
+
+
+@pytest.fixture(scope='module')
 def multichannel_stream():
     """ create a multichannel stream for tests """
     return generate_multichannel_stream()
@@ -290,8 +332,7 @@ stream_funcs = {fname + '_' + mname: corr.get_stream_xcorr(fname, mname)
                 if fname != 'default'}
 
 
-@pytest.fixture(scope='module')
-def stream_cc_output_dict(multichannel_templates, multichannel_stream):
+def _stream_cc_output_dict(multichannel_templates, multichannel_stream):
     """ return a dict of outputs from all stream_xcorr functions """
     # corr._get_array_dicts(multichannel_templates, multichannel_stream)
     out = {}
@@ -321,9 +362,32 @@ def stream_cc_output_dict(multichannel_templates, multichannel_stream):
 
 
 @pytest.fixture(scope='module')
+def stream_cc_output_dict(multichannel_templates, multichannel_stream):
+    """ return a dict of outputs from all stream_xcorr functions """
+    return _stream_cc_output_dict(
+        multichannel_templates=multichannel_templates,
+        multichannel_stream=multichannel_stream
+    )
+
+
+@pytest.fixture(scope='module')
+def stream_cc_weighted_output_dict(weighted_multichannel_templates, multichannel_stream):
+    return _stream_cc_output_dict(
+        multichannel_templates=weighted_multichannel_templates,
+        multichannel_stream=multichannel_stream
+    )
+
+
+@pytest.fixture(scope='module')
 def stream_cc_dict(stream_cc_output_dict):
     """ return just the cc arrays from the stream_cc functions """
     return {name: result[0] for name, result in stream_cc_output_dict.items()}
+
+
+@pytest.fixture(scope='module')
+def stream_cc_weighted_dict(stream_cc_weighted_output_dict):
+    """ return just the cc arrays from the stream_cc functions """
+    return {name: result[0] for name, result in stream_cc_weighted_output_dict.items()}
 
 
 @pytest.fixture(scope='module')
@@ -589,6 +653,21 @@ class TestArrayCorrelateFunctions:
                 np.save("cc2.npy", cc2)
             assert np.allclose(cc1, cc2, atol=self.atol)
 
+    def test_known_weight_application(self, array_ccs, array_ccs_weighted):
+        cc_names = list(array_ccs.keys())
+        print(cc_names)
+        for key1, key2 in itertools.combinations(cc_names, 2):
+            cc1 = array_ccs_weighted[key1]
+            cc2 = array_ccs_weighted[key2]
+            print(f"Comparing {key1} and {key2}")
+            assert np.allclose(cc1, cc2, atol=self.atol)
+        for cc_name in cc_names:
+            print(f"Checking for {cc_name}")
+            assert np.allclose(
+                array_ccs_weighted[cc_name],
+                array_ccs[cc_name] * GLOBAL_WEIGHT,
+                atol=1e-5)
+
     def test_autocorrelation(self, array_ccs):
         """ ensure an autocorrelation occurred in each of ccs where it is
         expected, defined by starting_index variable """
@@ -617,6 +696,17 @@ class TestStreamCorrelateFunctions:
         # get correlation results into a list
         cc_names = list(stream_cc_dict.keys())
         cc_list = [stream_cc_dict[cc_name] for cc_name in cc_names]
+        cc_1 = cc_list[0]
+        # loop over correlations and compare each with the first in the list
+        # this will ensure all cc are "close enough"
+        for cc_name, cc in zip(cc_names[2:], cc_list[2:]):
+            assert np.allclose(cc_1, cc, atol=self.atol)
+
+    def test_weighted_multi_channel_xcorr(self, stream_cc_weighted_dict):
+        """ test various correlation methods weighted with multiple channels """
+        # get correlation results into a list
+        cc_names = list(stream_cc_weighted_dict.keys())
+        cc_list = [stream_cc_weighted_dict[cc_name] for cc_name in cc_names]
         cc_1 = cc_list[0]
         # loop over correlations and compare each with the first in the list
         # this will ensure all cc are "close enough"
@@ -837,10 +927,11 @@ class TestGenericStreamXcorr:
         """ ensure a callable can be registered """
         small_count = {}
 
-        def some_callable(template_array, stream_array, pad_array):
+        def some_callable(template_array, stream_array, pad_array,
+                          weight_array):
             small_count['name'] = 1
             return corr.numpy_normxcorr(template_array, stream_array,
-                                        pad_array)
+                                        pad_array, weight_array)
 
         func = corr.get_stream_xcorr(some_callable)
         func(multichannel_templates, multichannel_stream)
@@ -883,7 +974,7 @@ class TestRegisterNormXcorrs:
     # helper functions
     def name_func_is_registered(self, func_name):
         """ return True if func is registered as a normxcorr func """
-        # Note: don not remove this fixture or bad things will happen
+        # Note: do not remove this fixture or bad things will happen
         name = func_name.__name__ if callable(func_name) else func_name
         return name in corr.XCOR_FUNCS