+++ eqcorrscan/utils/correlate.py
@@ -184,10 +184,11 @@ def pool_boy(Pool, traces, **kwargs):
 def _pool_normxcorr(templates, stream, stack, pool, func, *args, **kwargs):
     chans = [[] for _i in range(len(templates))]
     array_dict_tuple = _get_array_dicts(templates, stream, stack=stack)
-    stream_dict, template_dict, pad_dict, seed_ids = array_dict_tuple
+    stream_dict, template_dict, pad_dict, weight_dict, seed_ids = array_dict_tuple
     # get parameter iterator
-    params = ((template_dict[sid], stream_dict[sid], pad_dict[sid])
-              for sid in seed_ids)
+    params = (
+        (template_dict[sid], stream_dict[sid], pad_dict[sid], weight_dict[sid])
+        for sid in seed_ids)
     # get cc results and used chans into their own lists
     results = [pool.apply_async(func, param) for param in params]
     try:
@@ -236,7 +237,7 @@ def _general_serial(func):
         no_chans = np.zeros(len(templates))
         chans = [[] for _ in range(len(templates))]
         array_dict_tuple = _get_array_dicts(templates, stream, stack=stack)
-        stream_dict, template_dict, pad_dict, seed_ids = array_dict_tuple
+        stream_dict, template_dict, pad_dict, weight_dict, seed_ids = array_dict_tuple
         if stack:
             cccsums = np.zeros([len(templates),
                                 len(stream[0]) - len(templates[0][0]) + 1])
@@ -246,7 +247,8 @@ def _general_serial(func):
         for chan_no, seed_id in enumerate(seed_ids):
             tr_cc, tr_chans = func(template_dict[seed_id],
                                    stream_dict[seed_id],
-                                   pad_dict[seed_id])
+                                   pad_dict[seed_id],
+                                   weight_dict[seed_id])
             if stack:
                 cccsums = np.sum([cccsums, tr_cc], axis=0)
             else:
@@ -405,9 +407,9 @@ def get_array_xcorr(name_or_func=None):
 
 
 @register_array_xcorr('numpy')
-def numpy_normxcorr(templates, stream, pads, *args, **kwargs):
+def numpy_normxcorr(templates, stream, pads, weights=None, *args, **kwargs):
     """
-    Compute the normalized cross-correlation using numpy and bottleneck.
+    Compute normalized cross-correlations of multiple templates using numpy.
 
     :param templates: 2D Array of templates
     :type templates: np.ndarray
@@ -415,12 +417,20 @@ def numpy_normxcorr(templates, stream, pads, *args, **kwargs):
     :type stream: np.ndarray
     :param pads: List of ints of pad lengths in the same order as templates
     :type pads: list
+    :param weights: 1D Array of weights to match templates
+    :type weights: np.ndarray
 
     :return: np.ndarray of cross-correlations
     :return: np.ndarray channels used
     """
     import bottleneck
 
+    # Handle weights
+    n_templates = templates.shape[0]
+    if weights is None:
+        weights = np.ones(n_templates)
+    assert weights.shape[0] == templates.shape[0], "Weights required for all "\
+                                                   "templates"
     # Generate a template mask
     used_chans = ~np.isnan(templates).any(axis=1)
     # Currently have to use float64 as bottleneck runs into issues with other
@@ -454,6 +464,7 @@ def numpy_normxcorr(templates, stream, pads, *args, **kwargs):
 
     for i, pad in enumerate(pads):
         res[i] = np.append(res[i], np.zeros(pad))[pad:]
+    res *= weights.reshape(n_templates, 1)
     return res.astype(np.float32), used_chans
 
 
@@ -470,8 +481,8 @@ def _centered(arr, newshape):
 
 
 @register_array_xcorr('time_domain')
-def time_multi_normxcorr(templates, stream, pads, threaded=False, *args,
-                         **kwargs):
+def time_multi_normxcorr(templates, stream, pads, weights=None, threaded=False,
+                         *args, **kwargs):
     """
     Compute cross-correlations in the time-domain using C routine.
 
@@ -481,12 +492,20 @@ def time_multi_normxcorr(templates, stream, pads, threaded=False, *args,
     :type stream: np.ndarray
     :param pads: List of ints of pad lengths in the same order as templates
     :type pads: list
+    :param weights: 1D Array of weights to match templates
+    :type weights: np.ndarray
     :param threaded: Whether to use the threaded routine or not
     :type threaded: bool
 
     :return: np.ndarray of cross-correlations
     :return: np.ndarray channels used
     """
+    # Handle weights
+    n_templates = templates.shape[0]
+    if weights is None:
+        weights = np.ones(n_templates)
+    assert weights.shape[0] == templates.shape[0], "Weights required for all "\
+                                                   "templates"
     used_chans = ~np.isnan(templates).any(axis=1)
 
     utilslib = _load_cdll('libutils')
@@ -514,7 +533,6 @@ def time_multi_normxcorr(templates, stream, pads, threaded=False, *args,
     templates = templates.astype(np.float32) - templates_means
     stream = stream.astype(np.float32) - stream_mean
     template_len = templates.shape[1]
-    n_templates = templates.shape[0]
     image_len = stream.shape[0]
     ccc_length = image_len - template_len + 1
     assert ccc_length > 0, "Template must be shorter than stream"
@@ -532,11 +550,14 @@ def time_multi_normxcorr(templates, stream, pads, threaded=False, *args,
         ccc[i] = np.append(ccc[i], np.zeros(pads[i]))[pads[i]:]
     templates += templates_means
     stream += stream_mean
+    # Apply weights
+    ccc *= weights.reshape(n_templates, 1)
     return ccc, used_chans
 
 
 @register_array_xcorr('fftw', is_default=True)
-def fftw_normxcorr(templates, stream, pads, threaded=False, *args, **kwargs):
+def fftw_normxcorr(templates, stream, pads, weights=None, threaded=False,
+                   *args, **kwargs):
     """
     Normalised cross-correlation using the fftw library.
 
@@ -556,6 +577,8 @@ def fftw_normxcorr(templates, stream, pads, threaded=False, *args, **kwargs):
     :type stream: np.ndarray
     :param pads: List of ints of pad lengths in the same order as templates
     :type pads: list
+    :param weights: 1D Array of weights to match templates
+    :type weights: np.ndarray
     :param threaded:
         Whether to use the threaded routine or not - note openMP and python
         multiprocessing don't seem to play nice for this.
@@ -564,6 +587,12 @@ def fftw_normxcorr(templates, stream, pads, threaded=False, *args, **kwargs):
     :return: np.ndarray of cross-correlations
     :return: np.ndarray channels used
     """
+    # Handle weights
+    n_templates = templates.shape[0]
+    if weights is None:
+        weights = np.ones(n_templates)
+    assert weights.shape[0] == templates.shape[0], "Weights required for all "\
+                                                   "templates"
     utilslib = _load_cdll('libutils')
 
     argtypes = [
@@ -659,6 +688,8 @@ def fftw_normxcorr(templates, stream, pads, threaded=False, *args, **kwargs):
                 variance_warning[0]))
     # Remove variance correction
     stream /= multiplier
+    # Apply weights
+    ccc *= weights.reshape(n_templates, 1)
     return ccc, used_chans
 
 
@@ -693,7 +724,7 @@ def _time_threaded_normxcorr(templates, stream, stack=True, *args, **kwargs):
     no_chans = np.zeros(len(templates))
     chans = [[] for _ in range(len(templates))]
     array_dict_tuple = _get_array_dicts(templates, stream, stack=stack)
-    stream_dict, template_dict, pad_dict, seed_ids = array_dict_tuple
+    stream_dict, template_dict, pad_dict, weight_dict, seed_ids = array_dict_tuple
     ccc_length = max(
         len(stream[0]) - len(templates[0][0]) + 1,
         len(templates[0][0]) - len(stream[0]) + 1)
@@ -704,7 +735,7 @@ def _time_threaded_normxcorr(templates, stream, stack=True, *args, **kwargs):
     for chan_no, seed_id in enumerate(seed_ids):
         tr_cc, tr_chans = time_multi_normxcorr(
             template_dict[seed_id], stream_dict[seed_id], pad_dict[seed_id],
-            True)
+            weights=weight_dict[seed_id], threaded=True)
         if stack:
             cccsums = np.sum([cccsums, tr_cc], axis=0)
         else:
@@ -774,12 +805,13 @@ def _fftw_stream_xcorr(templates, stream, stack=True, *args, **kwargs):
 
     chans = [[] for _i in range(len(templates))]
     array_dict_tuple = _get_array_dicts(templates, stream, stack=stack)
-    stream_dict, template_dict, pad_dict, seed_ids = array_dict_tuple
+    stream_dict, template_dict, pad_dict, weight_dict, seed_ids = array_dict_tuple
     assert set(seed_ids)
     cccsums, tr_chans = fftw_multi_normxcorr(
         template_array=template_dict, stream_array=stream_dict,
         pad_array=pad_dict, seed_ids=seed_ids, cores_inner=num_cores_inner,
-        cores_outer=num_cores_outer, stack=stack, *args, **kwargs)
+        cores_outer=num_cores_outer, weight_array=weight_dict, 
+        stack=stack, *args, **kwargs)
     no_chans = np.sum(np.array(tr_chans).astype(int), axis=0)
     for seed_id, tr_chan in zip(seed_ids, tr_chans):
         for chan, state in zip(chans, tr_chan):
@@ -799,6 +831,7 @@ def fftw_multi_normxcorr(
     seed_ids,
     cores_inner,
     cores_outer,
+    weight_array=None,
     stack=True,
     *args,
     **kwargs
@@ -814,6 +847,8 @@ def fftw_multi_normxcorr(
     :param pad_array:
     :type seed_ids: list
     :param seed_ids:
+    :type weight_array: dict
+    :param weight_array:
 
     rtype: np.ndarray, list
     :return: 3D Array of cross-correlations and list of used channels.
@@ -834,6 +869,8 @@ def fftw_multi_normxcorr(
                                flags='C_CONTIGUOUS'),
         np.ctypeslib.ndpointer(dtype=np.intc,
                                flags='C_CONTIGUOUS'),
+        np.ctypeslib.ndpointer(dtype=np.float32,
+                               flags='C_CONTIGUOUS'),
         ctypes.c_int, ctypes.c_int,
         np.ctypeslib.ndpointer(dtype=np.intc,
                                flags='C_CONTIGUOUS'),
@@ -853,7 +890,8 @@ def fftw_multi_normxcorr(
         fft-length
         used channels (stacked as per templates)
         pad array (stacked as per templates)
-        num threads inner
+        weight array (stack as per templates)
+        num thread inner
         num threads outer
         variance warnings
         missed correlation warnings (usually due to gaps)
@@ -910,6 +948,8 @@ def fftw_multi_normxcorr(
     used_chans_np = np.ascontiguousarray(used_chans, dtype=np.intc)
     pad_array_np = np.ascontiguousarray(
         [pad_array[seed_id] for seed_id in seed_ids], dtype=np.intc)
+    weight_array_np = np.ascontiguousarray(
+        [weight_array[seed_id] for seed_id in seed_ids], dtype=np.float32)
     variance_warnings = np.ascontiguousarray(
         np.zeros(n_channels), dtype=np.intc)
     missed_correlations = np.ascontiguousarray(
@@ -918,7 +958,7 @@ def fftw_multi_normxcorr(
     # call C function
     ret = utilslib.multi_normxcorr_fftw(
         template_array, n_templates, template_len, n_channels, stream_array,
-        image_len, cccs, fft_len, used_chans_np, pad_array_np,
+        image_len, cccs, fft_len, used_chans_np, pad_array_np, weight_array_np,
         cores_inner, cores_outer, variance_warnings, missed_correlations,
         int(stack))
     if ret < 0:
@@ -985,7 +1025,8 @@ def _run_fmf_xcorr(template_arr, data_arr, weights, pads, arch, step=1):
 
 
 @register_array_xcorr("fmf")
-def fmf_xcorr(templates, stream, pads, arch="precise", *args, **kwargs):
+def fmf_xcorr(templates, stream, pads, weights=None, arch="precise",
+              *args, **kwargs):
     """
     Compute cross-correlations in the time-domain using the FMF routine.
 
@@ -995,6 +1036,8 @@ def fmf_xcorr(templates, stream, pads, arch="precise", *args, **kwargs):
     :type stream: np.ndarray
     :param pads: List of ints of pad lengths in the same order as templates
     :type pads: list
+    :param weights: 1D Array of weights to match templates
+    :type weights: np.ndarray
     :param arch: "gpu" or "precise" to run on GPU or CPU respectively
     type arch: str
 
@@ -1003,6 +1046,12 @@ def fmf_xcorr(templates, stream, pads, arch="precise", *args, **kwargs):
     """
     assert templates.ndim == 2, "Templates must be 2D"
     assert stream.ndim == 1, "Stream must be 1D"
+    # Handle weights
+    n_templates = templates.shape[0]
+    if weights is None:
+        weights = np.ones(n_templates)
+    assert weights.shape[0] == templates.shape[0], "Weights required for all "\
+                                                   "templates"
 
     used_chans = ~np.isnan(templates).any(axis=1)
 
@@ -1011,7 +1060,7 @@ def fmf_xcorr(templates, stream, pads, arch="precise", *args, **kwargs):
         template_arr=templates.reshape(
             (1, templates.shape[0], templates.shape[1])).swapaxes(0, 1),
         data_arr=stream.reshape((1, stream.shape[0])),
-        weights=np.ones((1, templates.shape[0])),
+        weights=np.array([weights]),
         pads=np.array([pads]),
         arch=arch.lower())
 
@@ -1077,7 +1126,7 @@ def _fmf_multi_xcorr(templates, stream, *args, **kwargs):
 
     chans = [[] for _i in range(len(templates))]
     array_dict_tuple = _get_array_dicts(templates, stream, stack=True)
-    stream_dict, template_dict, pad_dict, seed_ids = array_dict_tuple
+    stream_dict, template_dict, pad_dict, weight_dict, seed_ids = array_dict_tuple
     assert set(seed_ids)
 
     # Reshape templates into [templates x traces x time]
@@ -1088,7 +1137,8 @@ def _fmf_multi_xcorr(templates, stream, *args, **kwargs):
     # Moveouts should be [templates x traces]
     pads = np.array([pad_dict[seed_id] for seed_id in seed_ids]).swapaxes(0, 1)
     # Weights should be shaped like pads
-    weights = np.ones_like(pads)
+    weights = np.array([weight_dict[seed_id]
+                        for seed_id in seed_ids]).swapaxes(0, 1)
 
     cccsums = _run_fmf_xcorr(
         template_arr=t_arr, weights=weights, pads=pads,
@@ -1141,6 +1191,14 @@ def get_stream_xcorr(name_or_func=None, concurrency=None):
 # --------------------------- stream prep functions
 
 
+def _get_weight(tr, default_weight=1):
+    if not hasattr(tr.stats, "extra"):
+        return default_weight
+    if not hasattr(tr.stats.extra, "weight"):
+        return default_weight
+    return tr.stats.extra.weight
+
+
 def _get_array_dicts(templates, stream, stack, copy_streams=True):
     """ prepare templates and stream, return dicts """
     # Do some reshaping
@@ -1148,6 +1206,7 @@ def _get_array_dicts(templates, stream, stack, copy_streams=True):
     template_dict = {}
     stream_dict = {}
     pad_dict = {}
+    weight_dict = {}
     t_starts = []
 
     stream.sort(['network', 'station', 'location', 'channel'])
@@ -1164,6 +1223,10 @@ def _get_array_dicts(templates, stream, stack, copy_streams=True):
         temps_with_seed = [template.traces[i].data for template in templates]
         t_ar = np.array(temps_with_seed).astype(np.float32)
         template_dict.update({seed_id: t_ar})
+        # Get weights - stored in trace.stats.extra.weight
+        weights = np.array([_get_weight(template[i])
+                            for template in templates])
+        weight_dict.update({seed_id: weights})
         stream_channel = _stream_quick_select(stream, seed_id.split('_')[0])[0]
         # Normalize data to ensure no float overflow
         stream_data = stream_channel.data / (np.max(
@@ -1189,7 +1252,7 @@ def _get_array_dicts(templates, stream, stack, copy_streams=True):
             pad_list = [0 for _ in templates]
         pad_dict.update({seed_id: pad_list})
 
-    return stream_dict, template_dict, pad_dict, seed_ids
+    return stream_dict, template_dict, pad_dict, weight_dict, seed_ids
 
 
 # Remove fmf if it isn't installed