+++ eqcorrscan/utils/src/libutils.h
@@ -74,7 +74,8 @@ int running_mean_var(double*, double*, int*, float*, long, long);
 int normxcorr_fftw_main(float*, long, long, float*, long, int, int, float*, long,
                         float*, float*, float*, fftwf_complex*, fftwf_complex*,
                         fftwf_complex*, fftwf_plan, fftwf_plan, fftwf_plan,
-                        int*, int*, int, int*, int*, int);
+                        int*, int*, float*, int, int*, int*, int);
+
 
 int normxcorr_fftw_threaded(
     float*, long, long, float*, long, float*, long, int*, int*, int*, int*);
@@ -88,7 +89,7 @@ void free_fftw_arrays(
      fftw_complex**);
 
 int multi_normxcorr_fftw(
-    float*, long, long, long, float*, long, float*, long, int*, int*, int,
+    float*, long, long, long, float*, long, float*, long, int*, int*, float,* int,
     int, int*, int*, int);
 
 int normxcorr_fftw(float*, long, long, float*, long, float*, long, int*, int*, int*, int*);